move otaclient.grpc.utils to otaclient.utils
